HEY! This STINKS!
CHARACTERIZATION:
  while we seem to correctly reflect the *overflow* behaviour expected for binary
  subtraction (`SBC` with `C`==1/no borrow) this is *NOT* the correct behaviour
  for `CMP`/`CPX`/`CPY`!

  6502 documentation says the `V` flag is *not affected* by CMP!

{{{
CMP

Compare memory and accumulator A–M

This is a very useful instruction for comparing the accumulator contents to the
contents of a memory location.
The status register flags are set according to the result of a subtraction of
the memory contents from the accumulator.
The accumulator contents are preserved but the status register flags may be
used to cause branches depending on the values which were compared.

Processor Status after use
C (carry flag): set if A greater than or equal to M
Z (zero flag): set if A=M
I (interrupt disable): not affected
D (decimal mode flag): not affected
B (break command): not affected
V (overflow flag): not affected <---------------------------------------
N (negative flag): set if bit 7 of the result is set
}}}